+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v0.20.0-dev (2016-XX-XX)
+## v0.20.0 (2016-02-17)
 
 * Enhancements/backwards incompatible changes
   * Extra modules now have to register through `Brando.Registry.register(MyApp.Module)`. You can supply what you wish to register as options. `[:menu, :gettext]` are the defaults.